Lexeme overhaul, Scanner modifications (see extended description)
Lexemes are now essentially wrappers around their DFA. They no longer take in some consumed input, as the concept of Tokens now handles this. In this sense, these Lexemes are no longer the "classical" definition of a Lexeme, but the concept is still similar enough to warrant the name.

Additionally, Lexemes now have a stringify method, which (for each Lexeme type) defines how a certain input will stringify under this lexeme. This is for the --tokens compilation flag.

To reflect the updated Lexeme schema, Scanner has been changed. Scanner now takes in an additional constructor parameter: an array of Lexemes with which to Scan a file. The tokenize function no longer returns anything, but instead scans the input file and generates Tokens for it. It must be used before the new hasNextToken, getNextToken, and peekNextToken functions can be called.

To reflect all this, Main has been updated.

These changes are done in preparation for implementing the Parser.